Move video view properties to new file
