+++ Library/ViewModels/VideoViewModel.swift
@@ -258,34 +258,3 @@ public final class VideoViewModel: VideoViewModelInputs, VideoViewModelOutputs, 
   public var inputs: VideoViewModelInputs { return self }
   public var outputs: VideoViewModelOutputs { return self }
 }
-
-public protocol VideoViewProperties {
-  var video: (hls: String?, high: String)? { get }
-  var photoFull: String { get }
-}
-
-public protocol HasVideoViewProperties {
-  var videoViewProperties: VideoViewProperties { get }
-}
-
-public typealias VideoViewConfiguration = HasServiceProjectWebURL & HasVideoViewProperties &
-  ProjectAnalyticsProperties
-
-public struct VideoViewPropertiesBox: VideoViewProperties {
-  public let video: (hls: String?, high: String)?
-  public let photoFull: String
-
-  public init(video: (hls: String?, high: String)?, photoFull: String) {
-    self.video = video
-    self.photoFull = photoFull
-  }
-}
-
-extension Project: HasVideoViewProperties {
-  public var videoViewProperties: VideoViewProperties {
-    VideoViewPropertiesBox(
-      video: self.video.map { ($0.hls, $0.high) },
-      photoFull: self.photo.full
-    )
-  }
-}
